backend/requirements.txt
Flask==2.3.4
psycopg2-binary==2.9.8
Flask-Cors==3.1.1